Split webhook code file into files for each broad categories
