+++ pkg/artifactory/resource_artifactory_webhook.go
@@ -11,7 +11,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/customdiff"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 )
 
 func contains(s []string, str string) bool {
@@ -62,30 +61,6 @@ type WebhookEventFilter struct {
 	Criteria   interface{} `json:"criteria"`
 }
 
-type BaseWebhookCriteria struct {
-	IncludePatterns []string `json:"includePatterns"`
-	ExcludePatterns []string `json:"excludePatterns"`
-}
-
-type RepoWebhookCriteria struct {
-	BaseWebhookCriteria
-	AnyLocal  bool     `json:"anyLocal"`
-	AnyRemote bool     `json:"anyRemote"`
-	RepoKeys  []string `json:"repoKeys"`
-}
-
-type BuildWebhookCriteria struct {
-	BaseWebhookCriteria
-	AnyBuild       bool     `json:"anyBuild"`
-	SelectedBuilds []string `json:"selectedBuilds"`
-}
-
-type ReleaseBundleWebhookCriteria struct {
-	BaseWebhookCriteria
-	AnyReleaseBundle              bool     `json:"anyReleaseBundle"`
-	RegisteredReleaseBundlesNames []string `json:"registeredReleaseBundlesNames"`
-}
-
 type WebhookHandler struct {
 	HandlerType       string                    `json:"handler_type"`
 	Url               string                    `json:"url"`
@@ -104,196 +79,6 @@ const webhooksUrl = "/event/api/v1/subscriptions"
 const webhookUrl = webhooksUrl + "/{webhookKey}"
 
 func resourceArtifactoryWebhook(webhookType string) *schema.Resource {
-	var baseWebhookBaseSchema = map[string]*schema.Schema{
-		"key": {
-			Type:             schema.TypeString,
-			Required:         true,
-			ValidateDiagFunc: validation.ToDiagFunc(validation.All(validation.StringLenBetween(2, 200), validation.StringDoesNotContainAny(" "))),
-			Description:      "Key of webhook. Must be between 2 and 200 characters. Cannot contain spaces.",
-		},
-		"description": {
-			Type:             schema.TypeString,
-			Optional:         true,
-			ValidateDiagFunc: validation.ToDiagFunc(validation.StringLenBetween(0, 1000)),
-			Description:      "Description of webhook. Max length 1000 characters.",
-		},
-		"enabled": {
-			Type:        schema.TypeBool,
-			Optional:    true,
-			Default:     true,
-			Description: "Status of webhook. Default to 'true'",
-		},
-		"event_types": {
-			Type:        schema.TypeSet,
-			Required:    true,
-			MinItems:    1,
-			Elem:        &schema.Schema{Type: schema.TypeString},
-			Description: "List of Events in Artifactory, Distribution, Release Bundle that function as the event trigger for the Webhook.",
-		},
-		"url": {
-			Type:             schema.TypeString,
-			Required:         true,
-			ValidateDiagFunc: validation.ToDiagFunc(validation.All(validation.IsURLWithHTTPorHTTPS, validation.StringIsNotEmpty)),
-			Description:      "Specifies the URL that the Webhook invokes. This will be the URL that Artifactory will send an HTTP POST request to.",
-		},
-		"secret": {
-			Type:             schema.TypeString,
-			Optional:         true,
-			ValidateDiagFunc: validation.ToDiagFunc(validation.StringIsNotEmpty),
-			Description:      "Secret authentication token that will be sent to the configured URL.",
-		},
-		"proxy": {
-			Type:             schema.TypeString,
-			Optional:         true,
-			ValidateDiagFunc: validation.ToDiagFunc(validation.StringIsNotEmpty),
-			Description:      "Proxy key from Artifactory Proxies setting",
-		},
-		"custom_http_headers": {
-			Type:        schema.TypeMap,
-			Optional:    true,
-			Elem:        &schema.Schema{Type: schema.TypeString},
-			Description: "Custom HTTP headers you wish to use to invoke the Webhook, comprise of key/value pair.",
-		},
-	}
-
-	var baseCriteriaSchema = map[string]*schema.Schema{
-		"include_patterns": {
-			Type:        schema.TypeSet,
-			Optional:    true,
-			Elem:        &schema.Schema{Type: schema.TypeString},
-			Description: `Simple comma separated wildcard patterns for repository artifact paths (with no leading slash).\nAnt-style path expressions are supported (*, **, ?).\nFor example: "org/apache/**"`,
-		},
-		"exclude_patterns": {
-			Type:        schema.TypeSet,
-			Optional:    true,
-			Elem:        &schema.Schema{Type: schema.TypeString},
-			Description: `Simple comma separated wildcard patterns for repository artifact paths (with no leading slash).\nAnt-style path expressions are supported (*, **, ?).\nFor example: "org/apache/**"`,
-		},
-	}
-
-	var repoWebhookSchema = mergeSchema(baseWebhookBaseSchema, map[string]*schema.Schema{
-		"criteria": {
-			Type:     schema.TypeSet,
-			Required: true,
-			MaxItems: 1,
-			Elem: &schema.Resource{
-				Schema: mergeSchema(baseCriteriaSchema, map[string]*schema.Schema{
-					"any_local": {
-						Type:        schema.TypeBool,
-						Required:    true,
-						Description: "Trigger on any local repositories",
-					},
-					"any_remote": {
-						Type:        schema.TypeBool,
-						Required:    true,
-						Description: "Trigger on any remote repositories",
-					},
-					"repo_keys": {
-						Type:        schema.TypeSet,
-						Required:    true,
-						Elem:        &schema.Schema{Type: schema.TypeString},
-						Description: "Trigger on this list of repository keys",
-					},
-				}),
-			},
-			Description: "Specifies where the webhook will be applied on which repositories.",
-		},
-	})
-
-	var buildWebhookSchema = mergeSchema(baseWebhookBaseSchema, map[string]*schema.Schema{
-		"criteria": {
-			Type:     schema.TypeSet,
-			Required: true,
-			MaxItems: 1,
-			Elem: &schema.Resource{
-				Schema: mergeSchema(baseCriteriaSchema, map[string]*schema.Schema{
-					"any_build": {
-						Type:        schema.TypeBool,
-						Required:    true,
-						Description: "Trigger on any builds",
-					},
-					"selected_builds": {
-						Type:        schema.TypeSet,
-						Required:    true,
-						Elem:        &schema.Schema{Type: schema.TypeString},
-						Description: "Trigger on this list of build IDs",
-					},
-				}),
-			},
-			Description: "Specifies where the webhook will be applied on which builds.",
-		},
-	})
-
-	var releaseBundleWebhookSchema = mergeSchema(baseWebhookBaseSchema, map[string]*schema.Schema{
-		"criteria": {
-			Type:     schema.TypeSet,
-			Required: true,
-			MaxItems: 1,
-			Elem: &schema.Resource{
-				Schema: mergeSchema(baseCriteriaSchema, map[string]*schema.Schema{
-					"any_release_bundle": {
-						Type:        schema.TypeBool,
-						Required:    true,
-						Description: "Trigger on any release bundles or distributions",
-					},
-					"registered_release_bundle_names": {
-						Type:        schema.TypeSet,
-						Required:    true,
-						Elem:        &schema.Schema{Type: schema.TypeString},
-						Description: "Trigger on this list of release bundle names",
-					},
-				}),
-			},
-			Description: "Specifies where the webhook will be applied, on which release bundles or distributions.",
-		},
-	})
-
-	var packRepoCriteria = func(artifactoryCriteria map[string]interface{}) map[string]interface{} {
-		return map[string]interface{}{
-			"any_local":  artifactoryCriteria["anyLocal"].(bool),
-			"any_remote": artifactoryCriteria["anyRemote"].(bool),
-			"repo_keys":  schema.NewSet(schema.HashString, artifactoryCriteria["repoKeys"].([]interface{})),
-		}
-	}
-
-	var packBuildCriteria = func(artifactoryCriteria map[string]interface{}) map[string]interface{} {
-		return map[string]interface{}{
-			"any_build":       artifactoryCriteria["anyBuild"].(bool),
-			"selected_builds": schema.NewSet(schema.HashString, artifactoryCriteria["selectedBuilds"].([]interface{})),
-		}
-	}
-
-	var packReleaseBundleCriteria = func(artifactoryCriteria map[string]interface{}) map[string]interface{} {
-		return map[string]interface{}{
-			"any_release_bundle":              artifactoryCriteria["anyReleaseBundle"].(bool),
-			"registered_release_bundle_names": schema.NewSet(schema.HashString, artifactoryCriteria["registeredReleaseBundlesNames"].([]interface{})),
-		}
-	}
-
-	var unpackRepoCriteria = func(terraformCriteria map[string]interface{}, baseCriteria BaseWebhookCriteria) interface{} {
-		return RepoWebhookCriteria{
-			AnyLocal:            terraformCriteria["any_local"].(bool),
-			AnyRemote:           terraformCriteria["any_remote"].(bool),
-			RepoKeys:            castToStringArr(terraformCriteria["repo_keys"].(*schema.Set).List()),
-			BaseWebhookCriteria: baseCriteria,
-		}
-	}
-
-	var unpackBuildCriteria = func(terraformCriteria map[string]interface{}, baseCriteria BaseWebhookCriteria) interface{} {
-		return BuildWebhookCriteria{
-			AnyBuild:            terraformCriteria["any_build"].(bool),
-			SelectedBuilds:      castToStringArr(terraformCriteria["selected_builds"].(*schema.Set).List()),
-			BaseWebhookCriteria: baseCriteria,
-		}
-	}
-
-	var unpackReleaseBundleCriteria = func(terraformCriteria map[string]interface{}, baseCriteria BaseWebhookCriteria) interface{} {
-		return ReleaseBundleWebhookCriteria{
-			AnyReleaseBundle:              terraformCriteria["any_release_bundle"].(bool),
-			RegisteredReleaseBundlesNames: castToStringArr(terraformCriteria["registered_release_bundle_names"].(*schema.Set).List()),
-			BaseWebhookCriteria:           baseCriteria,
-		}
-	}
 
 	var domainCriteriaLookup = map[string]interface{}{
 		"artifact":                   RepoWebhookCriteria{},
@@ -527,46 +312,6 @@ func resourceArtifactoryWebhook(webhookType string) *schema.Resource {
 		if err != nil && resp.StatusCode() == http.StatusNotFound {
 			data.SetId("")
 			return diag.FromErr(err)
-		}
-
-		return nil
-	}
-
-	var repoCriteriaValidation = func(criteria map[string]interface{}) error {
-		log.Print("[DEBUG] repoCriteriaValidation")
-
-		anyLocal := criteria["any_local"].(bool)
-		anyRemote := criteria["any_remote"].(bool)
-		repoKeys := criteria["repo_keys"].(*schema.Set).List()
-
-		if (anyLocal == false && anyRemote == false) && len(repoKeys) == 0 {
-			return fmt.Errorf("repo_keys cannot be empty when both any_local and any_remote are false")
-		}
-
-		return nil
-	}
-
-	var buildCriteriaValidation = func(criteria map[string]interface{}) error {
-		log.Print("[DEBUG] buildCriteriaValidation")
-
-		anyBuild := criteria["any_build"].(bool)
-		selectedBuilds := criteria["selected_builds"].(*schema.Set).List()
-
-		if anyBuild == false && len(selectedBuilds) == 0 {
-			return fmt.Errorf("selected_builds cannot be empty when any_build is false")
-		}
-
-		return nil
-	}
-
-	var releaseBundleCriteriaValidation = func(criteria map[string]interface{}) error {
-		log.Print("[DEBUG] releaseBundleCriteriaValidation")
-
-		anyReleaseBundle := criteria["any_release_bundle"].(bool)
-		registeredReleaseBundlesNames := criteria["registered_release_bundle_names"].(*schema.Set).List()
-
-		if anyReleaseBundle == false && len(registeredReleaseBundlesNames) == 0 {
-			return fmt.Errorf("registered_release_bundle_names cannot be empty when any_release_bundle is false")
 		}
 
 		return nil
